+++ newclient/src/modules/search/handlers/SearchDetails.vue
@@ -6,7 +6,6 @@
         :select-items="selectItems"
         :has-error="hasError"
         @onSelect="onSelect"
-        @getToken="getToken"
         @getAllSearch="getAllSearch"
         @routeTo="routeTo"
     />
@@ -57,7 +56,7 @@ export default class SearchDetails extends Vue {
      * Checks whether param is valid
      * @param param {Any}
      */
-    isValid(param): boolean {
+    isValidHash(param): boolean {
         return eth.isValidHash(this.removeSpaces(param)) || eth.isValidAddress(this.removeSpaces(param))
     }
     /**
@@ -68,10 +67,6 @@ export default class SearchDetails extends Vue {
         let routeName = ''
         this.isLoading = true
         this.hasError = false
-        if (!this.isValid(param)) {
-            this.setError(param)
-            return
-        }
         this.$apollo
             .query({
                 query: getHashType,
@@ -162,9 +157,6 @@ export default class SearchDetails extends Vue {
      * @param param {Any}
      */
     getToken(param): void {
-        if (!this.onlyLetters(param)) {
-            return
-        }
         this.isLoading = true
         this.hasError = false
         this.$apollo
@@ -194,10 +186,13 @@ export default class SearchDetails extends Vue {
      * @param param {Any}
      */
     getAllSearch(param): void {
+        if (this.removeSpaces(param).length === 0) {
+            return
+        }
         if (param && param.contract) {
             this.routeToToken(param.contract)
         }
-        this.onlyLetters(param) ? this.getToken(param) : this.getHashType(param)
+        this.isValidHash(param) ? this.getHashType(param) : this.getToken(param)
     }
     /**
      * Get selected value and route user to token page
@@ -217,14 +212,6 @@ export default class SearchDetails extends Vue {
             return val.replace(/ /g, '')
         }
         return ''
-    }
-    /**
-     * Check if param onl contains string
-     * @param param {Any}
-     */
-    onlyLetters(param): boolean {
-        const value = this.removeSpaces(param)
-        return /^[a-zA-Z]+$/.test(value) ? true : false
     }
     /*
   ===================================================================================