+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Release v1.0.6
+### Bug
+- Prevent searching an empty string & remove onlyLetters [#944](https://github.com/EthVM/EthVM/pull/944)
 
 ### Devop